Fix the sending of response text.
Since combining the logic of both question and response fields into
`PreviewableField`, the state mapping for response fields has been
broken. As such, when responses were posted, their values were not
properly sent.

To better mirror the parallel construction in `PreviewableField`, this
changeset removes questionFormData and responseFormData in favor of a
unified formData, with all the requisite changes to actions and
reducers.

See #30.