Rework the way that LaTeX items are rendered.
In order to support arbittary markup from WeBWorK, we have to be able
to pass HTML directly to WeBWorK. This doesn't work with the previous
method of LaTeX parsing, since it required that the non-LaTeX chunks
appearing between LaTeX chunks be rendered within a single HTML
wrapper element. But this can't work when the LaTeX is nested inside
of another element, because open and close tags for the containing
element get improperly nested in the dummy `span`.

The only solution is to stop rendering LaTeX bits as their own
components, and instead building a single string for the content of the
FormattedProblem component.

Sanitization will have to handle server-side, via `wp_kses()`. More to
come in a subsequent changeset.

See #32.